requirements.txt
Flask==3.0.0
twilio==8.11.0
pymongo==4.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2